selenium tests code re-factor.
